fix clang warnings

+++ source/newtonian/two_dimensional/ghost_point_generators/FreeFlowGhostGenerator.cpp
@@ -36,8 +36,8 @@ std::pair<ComputationalCell, ComputationalCell> FreeFlowGenerator::GetGhostGradi
 	res.second.pressure *= -1;
 	for (size_t i = 0; i < res.first.tracers.size(); ++i)
 	{
-		(res.first.tracers.begin() + i)->second *= -1;
-		(res.second.tracers.begin() + i)->second *= -1;
+	  (res.first.tracers.begin() + static_cast<int>(i))->second *= -1;
+	  (res.second.tracers.begin() + static_cast<int>(i))->second *= -1;
 	}
 	Vector2D normal = edge.neighbors.first == static_cast<int>(ghost_index) ? 
 		tess.GetMeshPoint(edge.neighbors.first) - tess.GetMeshPoint(edge.neighbors.second)
